+++ src/Context_ctx/index.d.ts
@@ -1,4 +1,8 @@
 import type { Ctx } from '@ctx-core/object'
 import type { Context } from 'solid-js'
-export declare const Context_ctx:Context<Ctx>
-export declare function use_Context_ctx():Ctx
+export declare const ctx__Context:Context<Ctx>
+export declare function ctx__Context__use():Ctx
+export {
+	ctx__Context__use as use_Context_ctx,
+	ctx__Context as Context_ctx
+}