+++ src/Context_ctx/index.js
@@ -1,5 +1,9 @@
 import { createContext, useContext } from 'solid-js'
-export const Context_ctx = createContext()
-export function use_Context_ctx() {
-	return useContext(Context_ctx)
+export const ctx__Context = createContext()
+export function ctx__Context__use() {
+	return useContext(ctx__Context)
+}
+export {
+	ctx__Context__use as use_Context_ctx,
+	ctx__Context as Context_ctx
 }